server: Move server testing knobs to their own file
Release justification: no functional change

Release note: None
